ai_data_analyzer/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
google-generativeai
python-dotenv
streamlit_tags
plotly
streamlit-toggle
xgboost
lightgbm